requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.27.1
pydantic==2.8.2
pydantic-settings==2.3.4
pandas==2.2.0
numpy==1.26.3
numexpr==2.8.8
polars==0.20.19
scikit-learn==1.4.0
pyodbc==5.1.0
psycopg[binary]==3.1.12
psutil==5.9.8
joblib==1.3.2
hnswlib==0.8.0
python-dotenv==1.0.1
requests==2.31.0
httpx>=0.27.0,<0.28.0  # Starlette 0.27.0 compatibility (TestClient)
SQLAlchemy==2.0.25
argon2-cffi==23.1.0
PyJWT==2.8.0
pywin32==306  # Outlook Desktop integration for email notifications
openpyxl==3.1.2  # Excel file generation (XLSX)

cachetools==5.3.2

# ============================================================================
# Testing & Development Dependencies
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
# httpx version constraint moved to main dependencies section
